Rename ingester to statsprocessor
